sstable: inject suffix replacement into block property filtering
This patch  teaches block property filters to update the passed in block
property at read time via the new `SyntheticSuffixIntersects()` interface
method. The BlockPropertyFilterer will only call this method if the underlying
reader was initialized with a synthetic suffix. This patch assumes that the
underlying sst with a suffix replacement rule only contains point keys,
implying the only BlockPropertyFilter implementation that needs to implement
this new method is the BlockPropertyIntervalFilter. When the filterer calls
`SyntheticIntersects` with a `BlockPropertyIntervalFilter`, the interval bounds
of the passed in block prop may update, depending on the new
`BlockIntervalSyntheticReplacer` that the `BlockPropertyIntervalFilter` is
initialized with. The production `BlockIntervalSyntheticReplacer` will live in
the cockroach repository, as it's aware of the cockroach key encoding.